requirements.txt
arrow==1.1.1
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
decorator==5.0.9
Django==3.2.6
django-rq==2.4.1
djangorestframework==3.12.4
Flask==2.0.1
httpie==2.4.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pygments==2.9.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-toolbelt==0.9.1
rq==1.9.0
rq-dashboard==0.6.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
validators==0.18.2
Werkzeug==2.0.1
